Major changes
Change the structure of the components and their CSS styles. Split the app into two parts: client and server. The client is the pre existing part of the app. The server is just for holding hardcoded data at the moment (may be changed at later time). Fetched furniture from the server so it renders dynamically through API.