Added and separated concerns
Made the detail page so it changes based on the product it should show. Added services folder that will hold API and other requests. userService file for user authentication and register functionality. Made small changes to the data on the server concerning the imageUrl property and some small CSS changes for more modern design.